Move `src/components/api` to `src/api`
