+++ src/api/quizQuestionUserAnswers.ts
@@ -1,12 +1,12 @@
 import {useMutation, useQuery} from 'react-query'
-import * as Api from '../../api'
-import queryClient from '../../queryClient'
+import * as Api from '../api'
+import queryClient from '../queryClient'
 import {
   QuestionCorrectness,
   QuizQuestionUserAnswer,
   QuizQuestionUserAnswerRequest,
   QuizRemoveAnswersRequest,
-} from '../../types'
+} from '../types'
 
 const P = 'quizQuestionUserAnswers'
