+++ src/components/Footer.tsx
@@ -16,7 +16,7 @@ import Flex from './core/Flex'
 import NextLink from './core/NextLink'
 import Text from './core/Text'
 import {useAuth} from '../AuthUserContext'
-import {useGetUser} from './api/user'
+import {useGetUser} from '../api/user'
 import {QueryGuard} from '../QueryGuard'
 import NewsletterForm from './domain/newsletter/NewsletterForm'
 import {routes} from '../routes'
