+++ src/components/domain/buttons/LogInOrOutButton.tsx
@@ -7,7 +7,7 @@ import {useAuth} from '../../../AuthUserContext'
 import {QueryGuard} from '../../../QueryGuard'
 import {routes} from '../../../routes'
 import {SocUser} from '../../../types'
-import {useGetUser} from '../../api/user'
+import {useGetUser} from '../../../api/user'
 import Button from '../../core/Button'
 import NextLink from '../../core/NextLink'
 import Loading from '../../Loading'
