+++ src/components/domain/course-review/CourseReviewItem.tsx
@@ -6,7 +6,7 @@ import Flex from '../../core/Flex'
 import Text from '../../core/Text'
 import Rating from '../../core/Rating'
 import EditCourseReview from './EditCourseReview'
-import {useDeleteCourseReview} from '../../api/courseReviews'
+import {useDeleteCourseReview} from '../../../api/courseReviews'
 import {useAuth} from '../../../AuthUserContext'
 import UserAvatar from '../user/UserAvatar'
 import Loading from '../../Loading'
