+++ src/components/domain/course-review/CourseReviews.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
-import {useGetCourseReviewsOverview} from '../../api/courseReviewOverviews'
-import {useGetCourseReviews} from '../../api/courseReviews'
+import {useGetCourseReviewsOverview} from '../../../api/courseReviewOverviews'
+import {useGetCourseReviews} from '../../../api/courseReviews'
 import {QueryGuard} from '../../../QueryGuard'
 import Heading from '../../core/Heading'
 import Text from '../../core/Text'
