+++ src/components/domain/lecture-comment/LectureCommentItem.tsx
@@ -4,7 +4,7 @@ import {useAuth} from '../../../AuthUserContext'
 import useEditItemActions from '../../../hooks/useEditItemActions'
 import {LectureComment} from '../../../types'
 import {formatDate} from '../../../utils'
-import {useDeleteLectureComment} from '../../api/lectureComments'
+import {useDeleteLectureComment} from '../../../api/lectureComments'
 import Flex from '../../core/Flex'
 import Text from '../../core/Text'
 import Loading from '../../Loading'
