+++ src/components/domain/lecture-comment/LectureComments.tsx
@@ -7,7 +7,7 @@ import Heading from '../../core/Heading'
 import AddLectureComment from './AddLectureComment'
 import LectureCommentItem from './LectureCommentItem'
 import {QueryGuard} from '../../../QueryGuard'
-import {useGetLectureComments} from '../../api/lectureComments'
+import {useGetLectureComments} from '../../../api/lectureComments'
 import {useAuth} from '../../../AuthUserContext'
 import Loading from '../../Loading'
 import {device} from '../../../theme/device'
