+++ src/components/domain/lecture/LectureQuiz.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import {QueryGuard} from '../../../QueryGuard'
 import {Lecture} from '../../../types'
-import {useGetQuizzesByLecture} from '../../api/quiz'
+import {useGetQuizzesByLecture} from '../../../api/quiz'
 import Quiz from './quiz/Quiz'
 
 type Props = {
