+++ src/components/landing-page/CoursesSlider.tsx
@@ -6,7 +6,7 @@ import {QueryGuard} from '../../QueryGuard'
 import {routes} from '../../routes'
 import {device} from '../../theme/device'
 import {CourseOverview} from '../../types'
-import {useGetCourses} from '../api/courses'
+import {useGetCourses} from '../../api/courses'
 import NextLink from '../core/NextLink'
 import CourseCard from '../domain/course/CourseCard'
 import Slider, {SliderTemplateProps} from '../Slider'
