+++ src/pages/kurzy/[slug]/index.tsx
@@ -25,7 +25,7 @@ import PageContentWrapper from '../../../components/PageContentWrapper'
 import NavBar from '../../../components/NavBar'
 import {useAuth} from '../../../AuthUserContext'
 import {UserAndQueryGuard} from '../../../QueryGuard'
-import {useGetCourseOverview} from '../../../components/api/courseOverview'
+import {useGetCourseOverview} from '../../../api/courseOverview'
 import CourseReviews from '../../../components/domain/course-review/CourseReviews'
 import NextLink from '../../../components/core/NextLink'
 import Loading from '../../../components/Loading'
