+++ src/pages/kurzy/index.tsx
@@ -10,7 +10,7 @@ import * as Api from '../../api'
 import PageContentWrapper from '../../components/PageContentWrapper'
 import NavBar from '../../components/NavBar'
 import {useAuth} from '../../AuthUserContext'
-import {useGetCourses} from '../../components/api/courses'
+import {useGetCourses} from '../../api/courses'
 import {UserAndQueryGuard} from '../../QueryGuard'
 import VoteNextCourse from '../../components/domain/vote/VoteNextCourse'
 import {device} from '../../theme/device'
