+++ src/pages/lektor/[slug].tsx
@@ -9,7 +9,7 @@ import Heading from '../../components/core/Heading'
 import Courses from '../../components/domain/course/Courses'
 import {GetStaticProps, GetStaticPropsContext, NextPage} from 'next'
 import {useAuth} from '../../AuthUserContext'
-import {useGetAuthorOverview} from '../../components/api/authorOverview'
+import {useGetAuthorOverview} from '../../api/authorOverview'
 import PageContentWrapper from '../../components/PageContentWrapper'
 import {useRouter} from 'next/router'
 import NavBar from '../../components/NavBar'
