+++ src/pages/onboarding.tsx
@@ -4,7 +4,7 @@ import {useRouter} from 'next/router'
 import React, {ChangeEvent, useState} from 'react'
 import styled from 'styled-components'
 import {useAuth} from '../AuthUserContext'
-import {useAddUser} from '../components/api/user'
+import {useAddUser} from '../api/user'
 import Button from '../components/core/Button'
 import CheckBox from '../components/core/CheckBox'
 import Flex from '../components/core/Flex'
