+++ src/pages/profil.tsx
@@ -10,10 +10,10 @@ import UserAvatar from '../components/domain/user/UserAvatar'
 import Courses from '../components/domain/course/Courses'
 import Button from '../components/core/Button'
 import TextField from '../components/core/TextField'
-import {useEditUser, useGetUser} from '../components/api/user'
+import {useEditUser, useGetUser} from '../api/user'
 import {NextPage} from 'next'
 import {useAuth} from '../AuthUserContext'
-import {useGetMyCourses} from '../components/api/myCourses'
+import {useGetMyCourses} from '../api/myCourses'
 import {useRouter} from 'next/router'
 import PageContentWrapper from '../components/PageContentWrapper'
 import NavBar from '../components/NavBar'
